\name{access_tutorial}
\alias{access_tutorial}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
A wrapper function to access the tutorial using the learnr package
}
\description{
A wrapper function to access the tutorial using the learnr package
}
\usage{
access_tutorial(tutorial)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{tutorial}{
A character value contianing the name of the tutorial.
}
}
\author{
Isaac Quintanilla Salinas
}
