% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getting_work.R
\name{getting_work}
\alias{getting_work}
\title{getting_work}
\usage{
getting_work(Tutorial, Name)
}
\arguments{
\item{Tutorial}{A character value containing the name of the tutorial.}

\item{Name}{A character value containing the name of the stundent.}
}
\description{
A function to move the csv file to your working directory.
}
