% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writing_tutorial.R
\name{writing_tutorial}
\alias{writing_tutorial}
\title{writing_tutorial Function}
\usage{
writing_tutorial(tutorial_id, tutorial_version, user_id, event, data)
}
\arguments{
\item{tutorial_id}{Item passed down from learnr recorder function.}

\item{tutorial_version}{Item passed down from learnr recorder function.}

\item{user_id}{Item passed down from learnr recorder function.}

\item{event}{Item passed down from learnr recorder function.}

\item{data}{Item passed down from learnr recorder function.}
}
\description{
Used to generate a csv file that will contain information needed for submission.
}
